requirements.txt
alabaster==0.7.12
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
commonmark==0.9.0
decorator==4.4.0
docutils==0.15.2
Flask==1.1.1
fluent==0.1.0
future==0.17.1
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
influxdb==5.2.3
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
llvmlite==0.30.0
MarkupSafe==1.1.1
more-itertools==7.2.0
numba==0.46.0
numpy==1.17.2
packaging==19.2
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.2
pytest==5.2.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
snowballstemmer==1.9.1
recommonmark==0.6.0
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
toml==0.10.0
torndb==0.3
traitlets==4.3.2
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
xlrd==1.2.0
zipp==0.6.0
